+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Added `C-arrow` support patch for `ranger`
+- Removed `.bak` as a hidden file in `ranger`
 - Use new `VSCode` telemetry setting as the old settings are deprecated
 - Added `pop-launcher`
 - Added `cosmic` module